Add a service to allow resetting PID errors.
This is useful when switching from velocity control to direct wheel
power control while the robot is in motion. Otherwise the error values
will remain when switching back to velocity control, which can result in
a jerking behaviour.

It is possible that the driver could be smarter and figure out when the
PID should be reset internally, however, doing it this way will avoid
breaking any existing code that could possibly depend on the current
behaviour.